requirements.txt
black
django
djangorestframework
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
lxml